Fix PR notes, create base class for stream integrations
